final int/long patches related to github #813

+++ core/src/main/java/org/bouncycastle/crypto/io/CipherInputStream.java
@@ -299,7 +299,7 @@ public class CipherInputStream
             int avail = available();
             if (n <= avail)
             {
-                bufOff += n;
+                bufOff += (int)n;
 
                 return n;
             }
